requirements.txt
google-genai>=0.3.0
google-api-core>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
xai-sdk>=0.1.0
tqdm>=4.0.0
